First version of create stream

+++ src/main/scala/org/alexisjago/videostream/stream/StreamService.scala
@@ -3,7 +3,7 @@ package org.alexisjago.videostream.stream
 trait StreamService {
   def getStreams(userId: Long): Int
 
-  def startStream(userId: Long): Int
+  def startStream(userId: Long): Long
 
   def stopStream(userId: Long, streamId: Long)
 }
